Security lint (#45)
Addressed instances of Implicit narrowing conversion in compound assignment flagged by security checking. Trivial since the issue was just a local int that should have been changed when MP, HP, etc. went to longs. Remaining instances seem to require changing global structures or method signatures. Keeping this small to validate that I know how to "commit changes with git".

* Update DiscoCombatHelper.java

* Implicit naaarrowing

Address instances of "Implicit narrowing conversion in compound assignment"

+++ src/net/sourceforge/kolmafia/combat/Macrofier.java
@@ -629,7 +629,7 @@ public class Macrofier
 
 	public static void macroCombo( StringBuffer macro, int[] combo )
 	{
-		int cost = 0;
+		long cost = 0;
 		for ( int i = 0; i < combo.length; ++i )
 		{
 			cost += SkillDatabase.getMPConsumptionById( combo[ i ] );
